Fix issue with mangled IDs.
MassPy seems to assume that metabolite IDs start with "M_" and that reaction IDs start with "R_"... but doesn't *require* them to.  But it'll strip those parts of the IDs if left unprompted.  But the wrapper was assuming that every ID originally had those characters, and was stripping them back off afterwards.  This change lets both be true.  (And adds a new omex test that doesn't have those characters in the IDs.)