Allow export of parameters.
MassPy seems to have a love-hate relationship with compartments, so those can't be exported, but parameters can.  No idea what actually triggers an element to be stuck in the 'boundary_conditions' list, since putting in a boundary species just puts it on the normal species concentrations list.  But if they put something there at some point, we'll be ready?  I guess?  At any rate, the new test checks for exporting a parameter and a boundary species.